refactored hooks directory
