+++ frontend/src/hooks/useLocationQuery/useLocationQuery.ts
